vAsm-2.0: Code emission: moved to single byte memory representation.
This fixes the code emission layer, to actually correspond to the 2.0 memory organization.

+++ include/BinaryVectors.h
@@ -14,6 +14,6 @@
 
 using namespace std;
 
-typedef vector<int32_t> Bloat;
+typedef vector<int8_t> Bloat;
 
 #endif /* BINARYVECTORS_H_ */
